feat(service) : 이미 예약된 테마 삭제 방지 기능 추가

+++ src/main/java/roomescape/service/ThemeService.java
@@ -4,15 +4,18 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.dto.ThemeRequestDto;
 import roomescape.model.Theme;
+import roomescape.repository.ReservedThemeChecker;
 import roomescape.repository.ThemeRepository;
 
 @Service
 public class ThemeService {
 
     private final ThemeRepository themeRepository;
+    private final ReservedThemeChecker reservedThemeChecker;
 
-    public ThemeService(ThemeRepository themeRepository) {
+    public ThemeService(ThemeRepository themeRepository, ReservedThemeChecker reservedThemeChecker) {
         this.themeRepository = themeRepository;
+        this.reservedThemeChecker = reservedThemeChecker;
     }
 
     public Theme addTheme(ThemeRequestDto themeRequestDto) {
@@ -26,6 +29,9 @@ public class ThemeService {
     }
 
     public void deleteTheme(Long id) {
+        if (reservedThemeChecker.isReservedTheme(id)) {
+            throw new IllegalArgumentException("Theme is already reserved.");
+        }
         themeRepository.deleteTheme(id);
     }
